requirements.txt
flask
flask-login
flask-wtf
flask-sqlalchemy
bcrypt
openai
python-dotenv
app
flask-migrate
bleach

openai>=1.3.0
PyMuPDF>=1.23.0
python-docx>=0.8.11